% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickhouse_connect.R
\name{clickhouse_connect}
\alias{clickhouse_connect}
\title{Establish a connection to ClickHouse}
\usage{
clickhouse_connect(
  dsn = NULL,
  host = config::get()$host,
  port = config::get()$port,
  database = config::get()$db,
  user = config::get()$user,
  password = config::get()$password
)
}
\arguments{
\item{dsn}{character(1) (optional) The Data Source Name (DSN) configured for
the ClickHouse ODBC driver. Created via the \verb{/home/user/.odbc.ini} file.}

\item{host}{The hostname or IP address of the ClickHouse server (optional if
DSN is used).}

\item{port}{The port of the ClickHouse server (default: 8443).}

\item{database}{The name of the ClickHouse database (default: "default").}

\item{user}{The username for authentication (default: "default").}

\item{password}{The password for authentication (default: "").}
}
\value{
A database connection object.
}
\description{
This function establishes a connection to a ClickHouse database using the
ODBC driver.
}
\details{
The \code{.odbc.ini} file is a configuration file that stores the connection
details for the ClickHouse ODBC driver. The file is located in the user's
home directory (\verb{/home/user/.odbc.ini}). The file should have the following
format:

\if{html}{\out{<div class="sourceCode">}}\preformatted{[ClickHouseDSN]
Driver=ClickHouse
Description=ClickHouse DSN
Server=localhost
Port=1234
Database=default
UID=default
PWD=foo
}\if{html}{\out{</div>}}

The input \code{dsn} should be the name of the DSN in the \code{.odbc.ini} file
which is \code{ClickHouseDSN} in the example above.

Note that a configuration file may also be used instead of an \code{.odbc.ini}
file. The configuration file should be located in the user's home directory
(\verb{/home/user/config.yml}). The file should have the following format:

\if{html}{\out{<div class="sourceCode">}}\preformatted{default:
  host: "localhost"
  user: "default"
  port: 8123
  password: "foo"
  db: "default"
}\if{html}{\out{</div>}}
}
\examples{
clickhouse_connect(
    host = "localhost", port = 8123, database = "default", user = "default"
)

}
