% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdClick.R
\name{cmdListTables}
\alias{cmdListTables}
\alias{cMDclick}
\alias{cmdColNames}
\alias{cmdHead}
\alias{cmdGetMarker}
\alias{cmdGetRelab}
\title{Helper functions to interact with ClickHouse using the ODBC driver.}
\usage{
cmdListTables(con)

cmdColNames(con, tblName)

cmdHead(con, tblName, n = 6)

cmdGetMarker(
  con,
  features = NULL,
  samples = NULL,
  type = c("abundance", "presence")
)

cmdGetRelab(
  con,
  features = NULL,
  samples = NULL,
  filter_by = c("ncbi_tax_id", "marker_concatenated")
)
}
\value{
\itemize{
\item cmdListTables() returns a character vector of table names
\item cmdColNames() returns a character vector of column names
\item cmdHead() returns a tibble of the first few rows of a table
\item cmdGetMarker() returns a tibble of marker data
\item cmdGetRelab() returns a tibble of bug relative abundance data
}
}
\description{
These functions provide a simple interface to interact with a
ClickHouse database using the ODBC driver.
}
\examples{
conn <- clickhouse_connect(dsn = "ClickHouseDSN")
cmdListTables(conn)
cmdColNames(conn, "src_cmgd_v4__marker_abundance")
cmdHead(conn, "src_cmgd_v4__marker_abundance")
cmdGetMarker(conn, samples = "SAMEA103958109")

cmdGetMarker(conn, features = "UniClust90_AGMDBMFK00910|1__15|SGB72336")
cmdGetRelab(
    conn, features = "k__Bacteria", filter_by = "marker_concatenated"
)
}
