% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMDClick.R
\name{cmdGetMarker}
\alias{cmdGetMarker}
\title{Get marker data}
\usage{
cmdGetMarker(
  con,
  features = NULL,
  samples = NULL,
  type = c("abundance", "presence")
)
}
\arguments{
\item{con}{A connection created with DBI.}

\item{features}{A \code{character} vector with marker ids.}

\item{samples}{A \code{character} vector with sample names.}

\item{type}{A \code{character(1)} vector. One of two options: "\code{abundance}" or
"\code{presence}".}
}
\value{
A data.frame.
}
\description{
\code{cmdGetMarker} gets data related to marker abundance or presence.
}
\examples{

\dontrun{

con <- function() {
    DBI::dbConnect(
        drv = ClickHouseHTTP::ClickHouseHTTP(),
        port = 8443,
        https = TRUE,
        host = config::get()$host,
        user = config::get()$user,
        password = config::get()$password,
        dbname = config::get()$db
    )
}

## Get selected markers abundance data
mAb <- cmdGetMarker(con(), features = exMarkers)

## Get selected samples presence data 
mPr <- cmdGetMarker(con(), samples = exSamples, type = "p")

## exMarkers and exSamples are character vectors included in the package.

}

}
\seealso{
\code{\link[=cmdGetRelab]{cmdGetRelab()}}
}
