% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMDClick.R
\name{cmdGetRelab}
\alias{cmdGetRelab}
\title{Get relative abundance data}
\usage{
cmdGetRelab(
  con,
  features = NULL,
  samples = NULL,
  id_type = c("ncbi", "metaphlan")
)
}
\arguments{
\item{con}{A connection created with DBI.}

\item{features}{A \code{character} vector with marker ids.}

\item{samples}{A \code{character} vector with sample names.}

\item{id_type}{A \code{character(1)} vector. Choose in which id column perform
the filtering. Options: "ncbi" or "metaphlan"}
}
\value{
A data.frame.
}
\description{
\code{cmdGetRalab} Get relative bugs relative abundance.
}
\examples{

\dontrun{

con <- function() {
    DBI::dbConnect(
        drv = ClickHouseHTTP::ClickHouseHTTP(),
        port = 8443,
        https = TRUE,
        host = config::get()$host,
        user = config::get()$user,
        password = config::get()$password,
        dbname = config::get()$db
    )
}

## Get selected features relative abundance data
relabFeatures <- cmdGetRelab(con(), features = exTaxa)

## Get selected samples relative abundance data
relabSamples <- cmdGetRelab(con(), samples = exSamples)

## exTaxa and exSamples are character vectors included in the package.

}

}
\seealso{
\code{\link[=cmdGetMarker]{cmdGetMarker()}}
}
