% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_config.R
\name{config}
\alias{config}
\alias{has_config}
\alias{valid_config}
\title{Check and validate the user configuration file}
\usage{
has_config()

valid_config()
}
\value{
A logical vector indicating if the user has an ODBC configuration or
a general configuration file (\code{has_config}) or a valid \code{config.yml} file
(\code{valid_config}).
}
\description{
These are utitlities to check whether the user has an ODBC
configuration file or a general configuration file. These files are used to
connect via ODBC driver either with an \code{.odbc.ini} file or a general
configuration file \code{config.yml} as used by \code{config} R package. For details,
see \code{\link[=clickhouse_connect]{clickhouse_connect()}}. \code{valid_config} checks if the general
configuration file has the required fields for connecting to the Clickhouse
server.
}
\examples{
has_config()
valid_config()
}
