% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_config.R
\name{has_config}
\alias{has_config}
\title{Check if the user has a configuration file}
\usage{
has_config()
}
\value{
A logical vector indicating if the user has an ODBC configuration or
a general configuration file.
}
\description{
The function checks if the user has an ODBC configuration file
or a general configuration file. These files are used to connect via ODBC
driver either with an \code{.odbc.ini} file or a general configuration file
\code{config.yml}. For details, see \code{\link[=clickhouse_connect]{clickhouse_connect()}}.
}
\examples{
has_config()

}
