% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet_setup.R
\name{parquet_setup}
\alias{parquet_setup}
\alias{parquet_import}
\alias{cMDClickCache}
\title{Setup a connection to a DuckDB database with Parquet data}
\usage{
parquet_setup(
  db = "cMDClick.duckdb",
  dbdir = cMDClickCache(verbose = verbose),
  verbose = FALSE
)

parquet_import(
  con,
  dataType = c("metaphlan_bugs", "marker_abundances", "marker_presences"),
  verbose = FALSE
)

cMDClickCache(
  cache_dir = getOption("cMDClickCache", tools::R_user_dir("cMDClick", "cache")),
  verbose = TRUE,
  ask = interactive()
)
}
\arguments{
\item{db}{\code{character(1)} The name of the database file (default:
"cMDClick.duckdb").}

\item{dbdir}{\code{character(1)} The directory where the database file is stored
(default: \code{cMDClickCache()}).}

\item{verbose}{\code{logical(1)} Show informative messages in the console
(default: \code{FALSE}).}

\item{con}{\code{duckdb_connection} A connection object to a DuckDB database.}

\item{dataType}{\code{character()} The type of data to import. Can be one or more
of "metaphlan_bugs", "marker_abundances", or "marker_presences".}

\item{cache_dir}{\code{character(1)} The directory where the cache is stored.}

\item{ask}{\code{logical(1)} Ask the user to create the cache directory if it does
not exist (default: \code{interactive()}).}
}
\description{
This function creates a connection to a DuckDB database and
imports Parquet data from \url{https://store.cancerdatasci.org}.
}
\examples{
if (interactive()) {
    con <- parquet_setup()
    parquet_import(con, dataType = "metaphlan_bugs", verbose = TRUE)
    DBI::dbListTables(con)
}
}
